+++ tests/tests_geomstats/test_expectation_maximization.py
@@ -3,7 +3,8 @@
 import geomstats.backend as gs
 import geomstats.tests
 from geomstats.geometry.poincare_ball import PoincareBall
-from geomstats.learning.expectation_maximization import RiemannianEM
+from geomstats.learning.expectation_maximization import \
+    find_normalization_factor, RiemannianEM, find_variance_from_index
 from geomstats.learning.frechet_mean import FrechetMean
 
 TOLERANCE = 1e-3
@@ -110,17 +111,16 @@ class TestEM(geomstats.tests.TestCase):
     @geomstats.tests.np_and_pytorch_only
     def test_normalization_factor(self):
         """Test for Gaussian distribution normalization factor."""
-        variances_range,\
-            normalization_factor_var,\
-            phi_inv_var = \
-            self.metric.normalization_factor_init(gs.arange(
-                ZETA_LOWER_BOUND, ZETA_UPPER_BOUND, ZETA_STEP))
+        gmm = RiemannianEM(self.metric)
+        variances_range, normalization_factor_var, phi_inv_var = \
+            gmm.normalization_factor_init(
+                gs.arange(ZETA_LOWER_BOUND, ZETA_UPPER_BOUND, ZETA_STEP))
         self.assertAllClose(
             normalization_factor_var[4], 0.00291884, TOLERANCE)
         self.assertAllClose(phi_inv_var[3], 0.00562326, TOLERANCE)
 
         variances_test = gs.array([0.8, 1.2])
-        norm_factor_test = self.metric.find_normalization_factor(
+        norm_factor_test = find_normalization_factor(
             variances_test, variances_range, normalization_factor_var)
         norm_factor_verdict = gs.array([0.79577319, 2.3791778])
         self.assertAllClose(norm_factor_test, norm_factor_verdict, TOLERANCE)
@@ -135,7 +135,7 @@ class TestEM(geomstats.tests.TestCase):
         self.assertAllClose(
             norm_factor_gradient_test, norm_factor_gradient_verdict, TOLERANCE)
 
-        find_var_test = self.metric.find_variance_from_index(
+        find_var_test = find_variance_from_index(
             gs.array([0.5, 0.4, 0.3, 0.2]), variances_range, phi_inv_var)
         find_var_verdict = gs.array([0.481, 0.434, 0.378, 0.311])
         self.assertAllClose(find_var_test, find_var_verdict, TOLERANCE)
